Veronica/backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# Vector store + embeddings
chromadb>=0.4.0
langchain-community>=0.3.0
openai>=1.0.0

# Web search + scraping
httpx>=0.27.0
tavily-python>=0.5.0
apscheduler>=3.10.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=12.0

# Sandbox (optional, for plugin execution)
RestrictedPython>=6.0

# Voice
elevenlabs>=1.0.0

# Testing
pytest>=8.0
